feat: add refresh token authentication and clean repository files

+++ server/routes/auth_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, Response, HTTPException, Request
 from sqlalchemy.orm import Session
 from config.db import get_db
 from auth.auth import hash_password, verify_hash
-from utils.jwt_token import generate_access_token, verify_token, generate_refresh_token
+from utils.jwt_token import generate_access_token, generate_refresh_token, hash_token
 from models import User, RefreshToken
 from schemas.user import UserCreate, UserResponse, UserLogin
 from errors.auth_errors import PasswordHashingError
@@ -53,12 +53,14 @@ def sign_up(user: UserCreate , db: Session = Depends(get_db)):
 @auth_router.post("/sign-in")
 def sign_in(user:UserLogin, response:Response, db: Session = Depends(get_db)):
 
+    # check if user already exists
     existing_user = db.query(User).filter(User.email == user.email).first()
 
     if not existing_user:
         logger.info("User does not exist in the database")
-        raise HTTPException(status_code=404, detail="U$2b$12$7GkzL8..ser not found")
+        raise HTTPException(status_code=404, detail="User not found")
     
+    # verify password hash
     verify_hash(user.password, existing_user.password)
     
     payload_response = {
@@ -66,18 +68,22 @@ def sign_in(user:UserLogin, response:Response, db: Session = Depends(get_db)):
         "email" : existing_user.email
     }
     
+    # generate access tokens
     access_token = generate_access_token(payload_response)
 
     refresh_token = generate_refresh_token(payload_response)
 
+    # store hash of refresh token in the database
     token_record = RefreshToken(
         user_id = existing_user.id,
-        token_hash = hash_password(refresh_token),
+        token_hash = hash_token(refresh_token),
         expires_in = datetime.now(timezone.utc) + timedelta(days=30)
     )
 
     db.add(token_record)
     db.commit()
+
+    # add cookies for the browser
 
     response.set_cookie(
         key="refresh_token",
@@ -104,10 +110,33 @@ def sign_in(user:UserLogin, response:Response, db: Session = Depends(get_db)):
     }
 
 @auth_router.post("/sign-out")
-def sign_out(response: Response):
-    
+def sign_out(request: Request, response: Response,db: Session = Depends(get_db)):
+
+    refresh_token = request.cookies.get("refresh_token")
+
+    if not refresh_token:
+        logger.info("Refresh token missing or expired!")
+        raise HTTPException(status_code=404, detail="Missing token")
+
+    hashed_refresh_token = hash_token(refresh_token)
+
+    db.query(RefreshToken).filter(RefreshToken.token_hash == hashed_refresh_token).delete()
+
+    db.commit()
+
     response.delete_cookie(
-        key="access_token", secure=False, httponly=True, samesite="lax")
+        key="access_token",
+        httponly=True,
+        secure=(ENVIRONMENT == "production"),
+        samesite="lax"
+    )
+
+    response.delete_cookie(
+        key="refresh_token",
+        httponly=True,
+        secure=(ENVIRONMENT == "production"),
+        samesite="lax"
+    )
 
     return {
         "success": True,
